docs(README): refresh for modern day
To get this to work in Hephy v2.22.1, with a recent go, I had to do:

deis config:set https://github.com/heroku/heroku-buildpack-go#v148

I think this is a bug in our detect routines, I think it came up at the
last open roadmap. But for now, it could be a workaround that we just
document in the README, so people maybe don't have to trip over this...

chore(Godeps.json,README.md): latest available go

The latest buildpack from Heroku is actually heroku-buildpack-go#v148
with go 1.15.3, (it looks like buildpack v149 is due out soon as there
is already a changelog for it in heroku's master branch 4 days ago. It
carries 1.15.4, the latest go 1.15.5 is not available on Heroku yet...)